Reorganize utility so most code is in mache itself
